+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix missing `k` options when using the `classify` method of `MsdClassifier` (parameter `k` was ignored / not transferred to the underlying `msd` function).
+
 ## 1.0.3 (2023-07-20)
 
 - Don't round values if `precision` option is explicitly set to `null` (partially fixes [#28](https://github.com/riatelab/statsbreaks/issues/28)).